Robo Car
This code is based on C++. It is a car that runs with a bluetooth module. Other than the 4 wheels & DC motors, it has ultrasonic sensors. The specialty of the vehicle is to open its lid fixed with the servo motor. It can carry stuffs and move to desired places.